requirements.txt
# Pharma Sales Territory Optimizer - Requirements
# Install with: pip install -r requirements.txt

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Data Generation
faker>=20.0.0

# Snowflake
snowflake-connector-python>=3.0.0

# Machine Learning
scikit-learn>=1.3.0

# Optimization
pulp>=3.0.0

# Visualization
plotly>=5.18.0
kaleido>=1.0.0

# Dashboard
streamlit>=1.30.0
folium>=0.15.0
streamlit-folium>=0.18.0

# PySpark (optional - for Phase 3)
pyspark>=3.4.0
